+++ dags/dag_C_maintenance_02.py
@@ -1,3 +1,7 @@
+# -------------------------------------- #
+# Imports
+# -------------------------------------- #
+
 from datetime import datetime
 import logging
 import shutil
@@ -7,6 +11,10 @@ from airflow.operators.python import PythonOperator
 
 from dag import dag_maker
 
+# -------------------------------------- #
+# DAG
+# -------------------------------------- #
+
 default_args = {
     "owner": "airflow",
     'depends_on_past': False,
@@ -14,8 +22,26 @@ default_args = {
     'email_on_failure': True
 }
 
-MAX_LOG_DAYS = 1
+with DAG(
+    dag_id="Clean_old_logs_v01",
+    start_date=datetime(2023, 1, 1),
+    schedule_interval="00 00 * * *",
+    default_args=default_args,
+    max_active_runs=1,
+    catchup=False,
+    ) as dag:
+
+
+# -------------------------------------- #
+# Global variables
+# -------------------------------------- #
+
+MAX_LOG_DAYS = 2
 LOG_DIR = '/efs/airflow/logs/'
+
+# -------------------------------------- #
+# FUNCTIONS
+# -------------------------------------- #
 
 def find_old_logs():
     # Query old dag runs and build the log file paths to be deleted
@@ -38,7 +64,8 @@ def find_old_logs():
         
         for row in rows:
             delete_log_dir(row[0])
-            
+
+
 def delete_log_dir(log_dir):
     try:
         # Recursively delete the log directory and its log contents (e.g, 1.log, 2.log, etc)
@@ -47,15 +74,16 @@ def delete_log_dir(log_dir):
     except OSError as e:
         logging.info(f"Unable to delete: {e.filename} - {e.strerror}")
         
-        with DAG(
-            dag_id="airflow_log_cleanup",
-            start_date=datetime(2021, 1, 1),
-            schedule_interval="00 00 * * *",
-            default_args=default_args,
-            max_active_runs=1,
-            catchup=False,
-            ) as dag:
-                log_cleanup_op = PythonOperator(
-                    task_id="delete_old_logs",
-                    python_callable=find_old_logs
-                )
+
+# -------------------------------------- #
+# TASKS
+# -------------------------------------- #
+
+log_cleanup_op = PythonOperator(
+    task_id="delete_old_logs",
+    python_callable=find_old_logs
+    )
+
+# -------------------------------------- #
+# DEPENDANCIES
+# -------------------------------------- #
